backend/requirements.txt
# ============================================================================
# Python Version Requirement: >=3.11.9, <3.12
# ============================================================================
# CRITICAL: Python 3.12 has a breaking bug in async_generators that causes
# SSE (Server-Sent Events) streaming to fail with "async_generator already running"
# exceptions. This directly impacts our chat streaming functionality in Stage 1.
#
# The bug was introduced in CPython 3.12.0 and remains unfixed as of 3.12.7.
# Reference: https://github.com/python/cpython/issues/108668
#
# Python 3.11.9+ is required for:
# - Pydantic v2 performance improvements (30-50% faster validation)
# - FastAPI async improvements and better exception handling
# - SQLAlchemy 2.0 compatibility with improved typing support
#
# DO NOT upgrade to Python 3.12 until the async_generator bug is resolved.
# ============================================================================

# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17  # For file uploads (future use)

# Database ORM and migrations
sqlalchemy==2.0.36
alembic==1.14.0  # For database migrations (future use)

# Pydantic for validation
pydantic==2.10.0
pydantic-settings==2.6.1  # For Settings class

# HTTP client for LLM service
httpx==0.27.2  # Async HTTP client for llama.cpp

# SSE (Server-Sent Events) streaming
sse-starlette==1.6.5  # SSE support for FastAPI

# SECURITY FIX (SEC-003): CSRF Protection - Token-based implementation
# Provides robust CSRF protection using cryptographically signed tokens
# Tokens must be fetched from /api/csrf-token and included in X-CSRF-Token header
fastapi-csrf-protect==0.3.4

# Environment variables
python-dotenv==1.0.1

# CORS support (included in FastAPI but explicit for clarity)
# No separate package needed - FastAPI includes CORS middleware

# Logging and monitoring (Python stdlib, no packages needed)

# Testing dependencies (optional, for development)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Code quality tools (optional, for development)
ruff==0.8.2  # Linter and formatter
mypy==1.13.0  # Type checker
bandit==1.8.0  # Security linter

# SQLite driver (built into Python, no package needed)
# For PostgreSQL in future: psycopg2-binary==2.9.9
